moved label_images in resmaps
